Add read-module/port, use in mutation-runner tests to avoid files
By providing a way to read from a port, it's possible to write a simple function
to relocate a syntax object by changing the name of the port or the location
given to `read-syntax`.
This allows me to avoid doing a round trip through the file system to get module
syntax with the right location information.

However, ironically, I had to keep the round trips through the file system
to write some tests that helped me debug an issue with the paths being given to
the modules (which were coming out different than the paths on the syntax,
because the module reading code normalizes the module path).