Refactor mutant-factory to use process-Q, and make it resumable
This refactoring *significantly* simplifies the mutant-factory implementation.

It turns out using process-Q, mutant-factory can very easily be resumable, while
it would have been extremely difficult in the original implementation. The
original implementation was also huge, jumbled, and confusing, mostly because it
had an entire implicit process queue management system mixed in with the logic
of running the experiment. So that was Bad. It's better now.