Relax expression-selector assumptions on mutations with annotations
In particular, `select-exprs-as-if-untyped` assumed that no mutators will insert
or delete subexpressions from an expression containing type annotations.
However, we have just such a mutator: insert class method.

The reason for the assumption is that the selector must know where to put the
type annotations back into the syntax.

So this fix does a very simple possibly naive solution: it just changes the
syntax produced by the selector to explicitly group the sequences of
non-annotation exprs, then upon reconstructing removes those groupings.

Note that this could potentially prevent some mutations, but it's hard to know
how many. I doubt it's many.