Improve `make-instrumented-runner` exn handling
First, categorize exceptions that are raised by running the program into two
buckets:
1. Module evaluation errors
2. Runtime errors from running the program

Then, use these buckets to refine `run-with-mutated-module`s exn handling so
that compilation errors and runtime errors are not mixed.

This should help us catch early if there are any issues with transient in the
future, and also should improve our handling of unexpected errors arising when
running the program (e.g. the segfaults from TR unsoundness, or OS breaks).