modgraph: fix inference of fake dependencies between modules
It was getting confused when there are modules with names that are suffixes of
other modules. E.g.
A.rkt
another-A.rkt

If a module required just another-A.rkt, it would list both another-A.rkt and
A.rkt as dependencies, since searching for A.rkt in the program text gives a
hit.