mutant-factory: handle random unexpected type errors via revival
Sometimes some of our mutants produce a type error when they shouldn't be, with
an error from Typed Racket complaining about duplicate type annotations. This
happens apparently randomly with a low probability. After much investigation
it's still unclear what's causing this to happen, but this commit implements a
workaround: whenever a type error happens unexpectedly (ie the mutated module is
*not* typed), just re-run the mutant (do it up to 10 times).