Fix mutation changing source locations, causing wrong type err blame
Apparently somewhere in the mutation library I'm not being careful enough to
preserve source locations, and that was causing type errors to have the wrong
location in them (still within the program, rather than inside the mutation
library, so I'd guess it's from one of the id swap mutators). Instead of
imposing the constraint of preserving locations on the mutation library, just
replace the source locations of the entire mutated module after mutating.
For good measure, do this for all the unified modules too.

Also add a test that identifies the bug in mutation-runner.rkt