db-setup: scratch that, we need to check that type errors manifest
If we don't filter mutants for dynamic errors, then we'll have a bunch of
mutants where the mutation only affects a function that's never used, for
example, or similar.
Those aren't really interesting mistakes, so we need to filter them, and Natural
is the best we have that (theoretically) will signal an error if it is at all
observable dynamically.
Of course it will over-estimate, e.g. for unused argument types, but we've
already seen that doesn't happen much.