orchestration: improve db dir name management, and cut run-experiment.sh dependency
Before, the expected db set's directory needed to be configured across
1. experiment-info.rkt
2. experiment-manager.rkt
3. experiment configs
4. the run-experiment.sh script on the remote host

Which was ridiculous. Now, there's only two places:
1. experiment-info.rkt
2. experiment configs

Even these two could perhaps be collapsed... But that's for another time.

This solution has the added benefit that it also eliminates the external
dependency on the run-experiment.sh script existing and being up-to-date on the
remote host!