bex/orchestration: refactor db setup to track flows in seperate files
Before setup-all-dbs.rkt was the single *current* flow/process, and it had these
flags to tweak it but for a given experiment we really only wanted a particular
flow. So the process was just to edit and commit setup-all-dbs.rkt.

That is getting real confusing now that we've got multiple versions of the
experiment going at a time, and also made it pretty difficult to keep track of
the flow used in a given experiment -- especially since the underlying programs
changed over time, too!

Now, every version of the experiment we run can have its own file defining the
db setup process, so that we keep track of the differences and can keep them up
to date with changes to the underlying programs as necessary.