bex-data-analysis: improve DB data storage, use it for analysis
DB now contains serialized configs, nearly halving the space needed for a full
experiment's dataset.
It also now has metadata to support interpretation, analysis, and reproduction
of the data set.

So now everything is based around the results DB instead of the raw data files.

Unfortunately still need to have an experiment config installed to be able to
deserialize mutants, but it would be a good bit more work to remove that
dependency.